# histograms: Optimize query performance