--- a/promql/test.go
+++ b/promql/test.go
@@ -281,7 +281,7 @@ func (*evalCmd) testCmd()  {}
 type loadCmd struct {
 	gap       time.Duration
 	metrics   map[uint64]labels.Labels
-	defs      map[uint64][]Point
+	defs      map[uint64][]FPoint
 	exemplars map[uint64][]exemplar.Exemplar
 }
 
@@ -289,7 +289,7 @@ func newLoadCmd(gap time.Duration) *loadCmd {
 	return &loadCmd{
 		gap:       gap,
 		metrics:   map[uint64]labels.Labels{},
-		defs:      map[uint64][]Point{},
+		defs:      map[uint64][]FPoint{},
 		exemplars: map[uint64][]exemplar.Exemplar{},
 	}
 }
@@ -302,13 +302,13 @@ func (cmd loadCmd) String() string {
 func (cmd *loadCmd) set(m labels.Labels, vals ...parser.SequenceValue) {
 	h := m.Hash()
 
-	samples := make([]Point, 0, len(vals))
+	samples := make([]FPoint, 0, len(vals))
 	ts := testStartTime
 	for _, v := range vals {
 		if !v.Omitted {
-			samples = append(samples, Point{
+			samples = append(samples, FPoint{
 				T: ts.UnixNano() / int64(time.Millisecond/time.Nanosecond),
-				V: v.Value,
+				F: v.Value,
 			})
 		}
 		ts = ts.Add(cmd.gap)
@@ -323,7 +323,7 @@ func (cmd *loadCmd) append(a storage.Appender) error {
 		m := cmd.metrics[h]
 
 		for _, s := range smpls {
-			if _, err := a.Append(0, m, s.T, s.V); err != nil {
+			if _, err := a.Append(0, m, s.T, s.F); err != nil {
 				return err
 			}
 		}
@@ -399,8 +399,8 @@ func (ev *evalCmd) compareResult(result parser.Value) error {
 			if ev.ordered && exp.pos != pos+1 {
 				return fmt.Errorf("expected metric %s with %v at position %d but was at %d", v.Metric, exp.vals, exp.pos, pos+1)
 			}
-			if !almostEqual(exp.vals[0].Value, v.V) {
-				return fmt.Errorf("expected %v for %s but got %v", exp.vals[0].Value, v.Metric, v.V)
+			if !almostEqual(exp.vals[0].Value, v.F) {
+				return fmt.Errorf("expected %v for %s but got %v", exp.vals[0].Value, v.Metric, v.F)
 			}
 
 			seen[fp] = true
@@ -409,7 +409,7 @@ func (ev *evalCmd) compareResult(result parser.Value) error {
 			if !seen[fp] {
 				fmt.Println("vector result", len(val), ev.expr)
 				for _, ss := range val {
-					fmt.Println("    ", ss.Metric, ss.Point)
+					fmt.Println("    ", ss.Metric, ss.T, ss.F)
 				}
 				return fmt.Errorf("expected metric %s with %v not found", ev.metrics[fp], expVals)
 			}
@@ -576,15 +576,15 @@ func (t *Test) exec(tc testCommand) error {
 			mat := rangeRes.Value.(Matrix)
 			vec := make(Vector, 0, len(mat))
 			for _, series := range mat {
-				for _, point := range series.Points {
+				for _, point := range series.Floats {
 					if point.T == timeMilliseconds(iq.evalTime) {
-						vec = append(vec, Sample{Metric: series.Metric, Point: point})
+						vec = append(vec, Sample{Metric: series.Metric, T: point.T, F: point.F})
 						break
 					}
 				}
 			}
 			if _, ok := res.Value.(Scalar); ok {
-				err = cmd.compareResult(Scalar{V: vec[0].Point.V})
+				err = cmd.compareResult(Scalar{V: vec[0].F})
 			} else {
 				err = cmd.compareResult(vec)
 			}
@@ -763,7 +763,7 @@ func (ll *LazyLoader) appendTill(ts int64) error {
 				ll.loadCmd.defs[h] = smpls[i:]
 				break
 			}
-			if _, err := app.Append(0, m, s.T, s.V); err != nil {
+			if _, err := app.Append(0, m, s.T, s.F); err != nil {
 				return err
 			}
 			if i == len(smpls)-1 {