Based on memory profiling, this does two things (in separate commits):
1. Use separate `FPoint` and `HPoint` types instead of the one “polymorphous” `Point` type in the PromQL engine.
2. Use separate sample types in the `storage.sampleRing`.

The 2nd step can be implemented elegantly with interfaces, but the gain in memory allocation is then more or less compensated by the additional overhead of the interfaces. Thus, the 3rd commit circumvents the interfaces under the hood in the (very common) case that only one sample type is used within one sampleRing.

The individual commit descriptions list more details.

I will also add some benchmark results below.

The Prombench results didn't reflect the performance degradations we can see in the microbenchmarks. However, the latter might be relevant for more expensive queries than those run in Prombench.

In the best case, of course, this is not only a performance improvement, but also a gain in clarity. I would argue that this is at least partially the case for the 1st commit. Some code is even clearer now, and the behavior of functions not (yet) supporting native histograms is also more what the user would expect. (They now ignore histograms rather than treating them as floats of value zero.)

For the 2nd and 3rd commit, the code tends to get more convoluted. I would see those commits really as a performance optimization that is mostly under the hood, limited to the `storage` package. It also required fewer and more localized code changes than the 1st commit.

Note that with the separate `FPoint` and `HPoint` types, the habit of calling a sample with a float value just a "sample" and calling a float value itself just a "sample" got more and more into the way. A sample with a histogram is still a "sample", and a histogram value is still a "value". Therefore, we should distinguish both by calling them by the property that separates them, i.e. "float", "histogram", and "float histogram", or in short "F", "H", "FH". The first three commits do a bit of the renaming where it got directly into the way, but the fourth PR tries to cleanup all the other locations. Note that we cannot simply change the protobufs or the naming in the JSON APIs. We leave things as they are there, but we can still avoid confusion for the internal variable names.