@beorn7 2022-12-09
I won't list all the benchmarks here. I selected `RangeQuery/expr=rate(a_...[1d])` and `RangeQuery/expr=changes(a_...[1d])` as benchmarks that showed quite a bit of degradation despite not touching actual histogram samples at all.

First let's recap where we are:

This is the degradation of the latest release-2.39 state compared to the state before this PR (commit 6dd4e907a, i.e. more or less what's released in 2.40):

```
name                                                  old time/op    new time/op    delta
RangeQuery/expr=changes(a_one[1d]),steps=1-16            391µs ± 2%     542µs ± 1%  +38.58%  (p=0.000 n=9+8)
RangeQuery/expr=changes(a_one[1d]),steps=10-16           452µs ± 2%     617µs ± 2%  +36.48%  (p=0.000 n=10+10)
RangeQuery/expr=changes(a_one[1d]),steps=100-16         1.12ms ± 1%    1.36ms ± 2%  +21.58%  (p=0.000 n=8+10)
RangeQuery/expr=changes(a_one[1d]),steps=1000-16        7.83ms ± 1%    8.94ms ± 1%  +14.21%  (p=0.000 n=10+10)
RangeQuery/expr=changes(a_ten[1d]),steps=1-16           2.98ms ± 0%    3.30ms ± 1%  +10.67%  (p=0.000 n=9+10)
RangeQuery/expr=changes(a_ten[1d]),steps=10-16          3.66ms ± 1%    4.10ms ± 1%  +11.82%  (p=0.000 n=10+10)
RangeQuery/expr=changes(a_ten[1d]),steps=100-16         10.5ms ± 0%    11.8ms ± 1%  +12.50%  (p=0.000 n=8+10)
RangeQuery/expr=changes(a_ten[1d]),steps=1000-16        77.6ms ± 1%    87.4ms ± 1%  +12.63%  (p=0.000 n=9+9)
RangeQuery/expr=changes(a_hundred[1d]),steps=1-16       30.4ms ± 2%    32.8ms ± 1%   +8.01%  (p=0.000 n=10+10)
RangeQuery/expr=changes(a_hundred[1d]),steps=10-16      37.1ms ± 2%    40.6ms ± 2%   +9.64%  (p=0.000 n=10+10)
RangeQuery/expr=changes(a_hundred[1d]),steps=100-16      105ms ± 1%     117ms ± 1%  +11.69%  (p=0.000 n=10+10)
RangeQuery/expr=changes(a_hundred[1d]),steps=1000-16     783ms ± 3%     876ms ± 1%  +11.83%  (p=0.000 n=9+10)
RangeQuery/expr=rate(a_one[1d]),steps=1-16            383µs ± 3%     538µs ± 2%  +40.49%  (p=0.000 n=10+9)
RangeQuery/expr=rate(a_one[1d]),steps=10-16           425µs ± 1%     614µs ± 2%  +44.59%  (p=0.000 n=10+10)
RangeQuery/expr=rate(a_one[1d]),steps=100-16          946µs ± 1%    1259µs ± 1%  +33.12%  (p=0.000 n=10+10)
RangeQuery/expr=rate(a_one[1d]),steps=1000-16        6.81ms ± 2%    7.90ms ± 0%  +16.04%  (p=0.000 n=10+9)
RangeQuery/expr=rate(a_ten[1d]),steps=1-16           2.95ms ± 1%    3.32ms ± 2%  +12.20%  (p=0.000 n=10+10)
RangeQuery/expr=rate(a_ten[1d]),steps=10-16          3.61ms ± 1%    4.02ms ± 1%  +11.41%  (p=0.000 n=10+10)
RangeQuery/expr=rate(a_ten[1d]),steps=100-16         9.56ms ± 1%   10.74ms ± 1%  +12.29%  (p=0.000 n=10+9)
RangeQuery/expr=rate(a_ten[1d]),steps=1000-16        67.7ms ± 2%    77.8ms ± 1%  +15.02%  (p=0.000 n=10+8)
RangeQuery/expr=rate(a_hundred[1d]),steps=1-16       29.8ms ± 1%    32.8ms ± 1%  +10.05%  (p=0.000 n=10+8)
RangeQuery/expr=rate(a_hundred[1d]),steps=10-16      36.6ms ± 1%    40.0ms ± 3%   +9.26%  (p=0.000 n=9+10)
RangeQuery/expr=rate(a_hundred[1d]),steps=100-16     95.2ms ± 2%   107.1ms ± 1%  +12.59%  (p=0.000 n=10+9)
RangeQuery/expr=rate(a_hundred[1d]),steps=1000-16     675ms ± 2%     781ms ± 0%  +15.59%  (p=0.000 n=10+8)

name                                                  old alloc/op   new alloc/op   delta
RangeQuery/expr=changes(a_one[1d]),steps=1-16            607kB ± 0%    1137kB ± 0%  +87.39%  (p=0.000 n=10+10)
RangeQuery/expr=changes(a_one[1d]),steps=10-16           607kB ± 0%    1136kB ± 0%  +87.16%  (p=0.000 n=10+10)
RangeQuery/expr=changes(a_one[1d]),steps=100-16          607kB ± 0%    1127kB ± 0%  +85.57%  (p=0.000 n=10+10)
RangeQuery/expr=changes(a_one[1d]),steps=1000-16         611kB ± 0%    1137kB ± 1%  +86.18%  (p=0.000 n=8+10)
RangeQuery/expr=changes(a_ten[1d]),steps=1-16           1.09MB ± 0%    1.61MB ± 0%  +47.78%  (p=0.000 n=10+10)
RangeQuery/expr=changes(a_ten[1d]),steps=10-16          1.08MB ± 0%    1.61MB ± 0%  +47.98%  (p=0.000 n=10+9)
RangeQuery/expr=changes(a_ten[1d]),steps=100-16         1.09MB ± 1%    1.62MB ± 0%  +48.20%  (p=0.000 n=10+8)
RangeQuery/expr=changes(a_ten[1d]),steps=1000-16        1.16MB ± 2%    1.74MB ± 0%  +50.21%  (p=0.000 n=10+8)
RangeQuery/expr=changes(a_hundred[1d]),steps=1-16       5.93MB ± 0%    6.42MB ± 1%   +8.14%  (p=0.000 n=10+10)
RangeQuery/expr=changes(a_hundred[1d]),steps=10-16      5.93MB ± 1%    6.42MB ± 1%   +8.36%  (p=0.000 n=10+10)
RangeQuery/expr=changes(a_hundred[1d]),steps=100-16     6.03MB ± 0%    6.56MB ± 0%   +8.75%  (p=0.000 n=8+8)
RangeQuery/expr=changes(a_hundred[1d]),steps=1000-16    6.73MB ± 0%    7.24MB ± 4%   +7.49%  (p=0.000 n=7+10)
RangeQuery/expr=rate(a_one[1d]),steps=1-16            605kB ± 0%    1139kB ± 0%  +88.19%  (p=0.000 n=10+9)
RangeQuery/expr=rate(a_one[1d]),steps=10-16           605kB ± 0%    1137kB ± 0%  +88.06%  (p=0.000 n=9+10)
RangeQuery/expr=rate(a_one[1d]),steps=100-16          604kB ± 0%    1127kB ± 0%  +86.59%  (p=0.000 n=10+10)
RangeQuery/expr=rate(a_one[1d]),steps=1000-16         607kB ± 0%    1136kB ± 0%  +87.02%  (p=0.000 n=10+10)
RangeQuery/expr=rate(a_ten[1d]),steps=1-16           1.08MB ± 0%    1.61MB ± 0%  +49.11%  (p=0.000 n=10+10)
RangeQuery/expr=rate(a_ten[1d]),steps=10-16          1.08MB ± 0%    1.61MB ± 0%  +48.85%  (p=0.000 n=10+8)
RangeQuery/expr=rate(a_ten[1d]),steps=100-16         1.09MB ± 1%    1.62MB ± 1%  +49.06%  (p=0.000 n=9+10)
RangeQuery/expr=rate(a_ten[1d]),steps=1000-16        1.17MB ± 0%    1.74MB ± 0%  +48.61%  (p=0.000 n=8+8)
RangeQuery/expr=rate(a_hundred[1d]),steps=1-16       5.89MB ± 0%    6.46MB ± 0%   +9.69%  (p=0.000 n=9+8)
RangeQuery/expr=rate(a_hundred[1d]),steps=10-16      5.89MB ± 0%    6.46MB ± 0%   +9.78%  (p=0.000 n=10+10)
RangeQuery/expr=rate(a_hundred[1d]),steps=100-16     5.96MB ± 1%    6.56MB ± 1%   +9.98%  (p=0.000 n=10+10)
RangeQuery/expr=rate(a_hundred[1d]),steps=1000-16    6.59MB ± 2%    7.46MB ± 1%  +13.22%  (p=0.000 n=10+8)
```

I would like to point out that the dramatic increases in runtime only happen for very small absolute runtimes, where the difference doesn't matter that much. Benchmarks with multiple milliseconds per op only show runtime increases of less than 20%, meaning that the runtime increase was never alarmingly bad.

The increase in allocations (I only show the volume, not the number of allocs, because the latter doesn't really change a lot) is more concerning. The relative increase tends to be higher with smaller volume, but even the smaller volumes are significant (smallest to largest volume is one order of magnitude difference, while the runtime spans three orders of magnitude).

With this in mind, let's look at the new comparison (2.39 vs. this PR), with special attention towards the allocations:

```
name                                                  old time/op    new time/op    delta
RangeQuery/expr=changes(a_one[1d]),steps=1-16            391µs ± 2%     451µs ± 1%  +15.37%  (p=0.000 n=9+10)
RangeQuery/expr=changes(a_one[1d]),steps=10-16           452µs ± 2%     526µs ± 5%  +16.30%  (p=0.000 n=10+10)
RangeQuery/expr=changes(a_one[1d]),steps=100-16         1.12ms ± 1%    1.22ms ± 2%   +9.26%  (p=0.000 n=8+10)
RangeQuery/expr=changes(a_one[1d]),steps=1000-16        7.83ms ± 1%    8.11ms ± 1%   +3.62%  (p=0.000 n=10+10)
RangeQuery/expr=changes(a_ten[1d]),steps=1-16           2.98ms ± 0%    3.68ms ± 1%  +23.39%  (p=0.000 n=9+10)
RangeQuery/expr=changes(a_ten[1d]),steps=10-16          3.66ms ± 1%    4.40ms ± 2%  +19.99%  (p=0.000 n=10+10)
RangeQuery/expr=changes(a_ten[1d]),steps=100-16         10.5ms ± 0%    11.4ms ± 2%   +9.03%  (p=0.000 n=8+10)
RangeQuery/expr=changes(a_ten[1d]),steps=1000-16        77.6ms ± 1%    80.6ms ± 2%   +3.91%  (p=0.000 n=9+10)
RangeQuery/expr=changes(a_hundred[1d]),steps=1-16       30.4ms ± 2%    37.3ms ± 1%  +22.81%  (p=0.000 n=10+9)
RangeQuery/expr=changes(a_hundred[1d]),steps=10-16      37.1ms ± 2%    44.2ms ± 3%  +19.32%  (p=0.000 n=10+10)
RangeQuery/expr=changes(a_hundred[1d]),steps=100-16      105ms ± 1%     114ms ± 2%   +8.42%  (p=0.000 n=10+10)
RangeQuery/expr=changes(a_hundred[1d]),steps=1000-16     783ms ± 3%     793ms ± 1%   +1.24%  (p=0.008 n=9+9)
RangeQuery/expr=rate(a_one[1d]),steps=1-16            383µs ± 3%     442µs ± 1%  +15.48%  (p=0.000 n=10+10)
RangeQuery/expr=rate(a_one[1d]),steps=10-16           425µs ± 1%     487µs ± 0%  +14.67%  (p=0.000 n=10+10)
RangeQuery/expr=rate(a_one[1d]),steps=100-16          946µs ± 1%     961µs ± 2%   +1.56%  (p=0.000 n=10+9)
RangeQuery/expr=rate(a_one[1d]),steps=1000-16        6.81ms ± 2%    5.92ms ± 3%  -13.04%  (p=0.000 n=10+10)
RangeQuery/expr=rate(a_ten[1d]),steps=1-16           2.95ms ± 1%    3.66ms ± 3%  +23.76%  (p=0.000 n=10+10)
RangeQuery/expr=rate(a_ten[1d]),steps=10-16          3.61ms ± 1%    4.14ms ± 1%  +14.70%  (p=0.000 n=10+8)
RangeQuery/expr=rate(a_ten[1d]),steps=100-16         9.56ms ± 1%    9.16ms ± 1%   -4.22%  (p=0.000 n=10+8)
RangeQuery/expr=rate(a_ten[1d]),steps=1000-16        67.7ms ± 2%    59.4ms ± 3%  -12.23%  (p=0.000 n=10+10)
RangeQuery/expr=rate(a_hundred[1d]),steps=1-16       29.8ms ± 1%    36.7ms ± 1%  +23.11%  (p=0.000 n=10+9)
RangeQuery/expr=rate(a_hundred[1d]),steps=10-16      36.6ms ± 1%    41.6ms ± 1%  +13.48%  (p=0.000 n=9+10)
RangeQuery/expr=rate(a_hundred[1d]),steps=100-16     95.2ms ± 2%    90.4ms ± 1%   -4.98%  (p=0.000 n=10+9)
RangeQuery/expr=rate(a_hundred[1d]),steps=1000-16     675ms ± 2%     584ms ± 2%  -13.58%  (p=0.000 n=10+10)

name                                                  old alloc/op   new alloc/op   delta
RangeQuery/expr=changes(a_one[1d]),steps=1-16            607kB ± 0%     611kB ± 0%   +0.64%  (p=0.000 n=10+10)
RangeQuery/expr=changes(a_one[1d]),steps=10-16           607kB ± 0%     609kB ± 0%   +0.43%  (p=0.000 n=10+10)
RangeQuery/expr=changes(a_one[1d]),steps=100-16          607kB ± 0%     608kB ± 0%     ~     (p=0.053 n=10+9)
RangeQuery/expr=changes(a_one[1d]),steps=1000-16         611kB ± 0%     612kB ± 1%     ~     (p=0.445 n=8+10)
RangeQuery/expr=changes(a_ten[1d]),steps=1-16           1.09MB ± 0%    1.09MB ± 0%     ~     (p=0.912 n=10+10)
RangeQuery/expr=changes(a_ten[1d]),steps=10-16          1.08MB ± 0%    1.09MB ± 0%   +0.19%  (p=0.004 n=10+9)
RangeQuery/expr=changes(a_ten[1d]),steps=100-16         1.09MB ± 1%    1.10MB ± 1%     ~     (p=0.190 n=10+10)
RangeQuery/expr=changes(a_ten[1d]),steps=1000-16        1.16MB ± 2%    1.16MB ± 2%     ~     (p=0.100 n=10+10)
RangeQuery/expr=changes(a_hundred[1d]),steps=1-16       5.93MB ± 0%    5.94MB ± 0%   +0.15%  (p=0.043 n=10+7)
RangeQuery/expr=changes(a_hundred[1d]),steps=10-16      5.93MB ± 1%    5.94MB ± 1%     ~     (p=0.190 n=10+10)
RangeQuery/expr=changes(a_hundred[1d]),steps=100-16     6.03MB ± 0%    6.05MB ± 1%     ~     (p=0.068 n=8+10)
RangeQuery/expr=changes(a_hundred[1d]),steps=1000-16    6.73MB ± 0%    6.74MB ± 0%   +0.06%  (p=0.000 n=7+8)
RangeQuery/expr=rate(a_one[1d]),steps=1-16            605kB ± 0%     611kB ± 0%   +0.90%  (p=0.000 n=10+10)
RangeQuery/expr=rate(a_one[1d]),steps=10-16           605kB ± 0%     609kB ± 0%   +0.79%  (p=0.000 n=9+10)
RangeQuery/expr=rate(a_one[1d]),steps=100-16          604kB ± 0%     609kB ± 0%   +0.75%  (p=0.000 n=10+10)
RangeQuery/expr=rate(a_one[1d]),steps=1000-16         607kB ± 0%     613kB ± 1%   +0.86%  (p=0.000 n=10+10)
RangeQuery/expr=rate(a_ten[1d]),steps=1-16           1.08MB ± 0%    1.09MB ± 1%   +0.61%  (p=0.000 n=10+10)
RangeQuery/expr=rate(a_ten[1d]),steps=10-16          1.08MB ± 0%    1.09MB ± 0%   +0.46%  (p=0.000 n=10+10)
RangeQuery/expr=rate(a_ten[1d]),steps=100-16         1.09MB ± 1%    1.09MB ± 0%   +0.51%  (p=0.013 n=9+10)
RangeQuery/expr=rate(a_ten[1d]),steps=1000-16        1.17MB ± 0%    1.17MB ± 3%     ~     (p=0.633 n=8+10)
RangeQuery/expr=rate(a_hundred[1d]),steps=1-16       5.89MB ± 0%    5.93MB ± 0%   +0.77%  (p=0.000 n=9+10)
RangeQuery/expr=rate(a_hundred[1d]),steps=10-16      5.89MB ± 0%    5.94MB ± 0%   +0.91%  (p=0.000 n=10+9)
RangeQuery/expr=rate(a_hundred[1d]),steps=100-16     5.96MB ± 1%    6.02MB ± 0%   +0.91%  (p=0.001 n=10+6)
RangeQuery/expr=rate(a_hundred[1d]),steps=1000-16    6.59MB ± 2%    6.66MB ± 3%   +1.10%  (p=0.014 n=10+10)
```

While there is still a noticeable, but smaller, degradation in runtime, the allocations have essentially not changed at all.